+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 10.2.3
 -----------------
 - Remove dead link to centresis.org in index.php
+- Fix MySQL error TEXT column used in key specification without a key length in Fields.fnc.php & DisciplineForm.php
 
 Changes in 10.2.2
 -----------------